pyproject.toml
[project]
name = "garagem"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jaotarzan", email = "joaovssouza59@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
